Card faces are distinct from values assigned:
Previously, what was 'drawn' (i.e: printed and displayed) was an int from a deck (e.g: 10). This was fine for most cards, but of course, face-cards (Jack, Queen, King) are represented by something other than their value (e.g: J - 10). This update allows for faceup cards to be shown in a more reflective manner as they would in the game - although the Ace card hasn't become interchangeable with values of either 1 or 11 for certain contexts and is the subject of the next update.