Fully functioning cards + match mode
This is the first version that has fully working aces that are adaptable to the overall score (i.e: ace which will become 1 if the total score is greater than 10 - which was a lot more awkward to program than I initially anticipated when I started this project). Also, I made a match mode with a tally counter that lets a player keep on playing until they wish to finish, keeping track and giving an "Overall winner".